Update indexing process to use asyncio
This is a significant rewrite of the long running indexing process to
use asyncio as the concurrency model. In addition, better monitoring of
the underlying ActiveMQ connection has been added.

The reindexing and thesis deletion have been temporarily removed until
they can also be updated to use asyncio. Until then, the old version can
be used.